#171 Refactor: move mockUser to data folder, change path and check user instead use class navigate
