+++ javascript/big-practice/data/mockUser.js
