+++ javascript/big-practice/src/scripts/main.js
@@ -3,6 +3,8 @@ import NoteModel from './models/noteModel';
 import NoteController from './controllers/noteController';
 import AuthenController from './controllers/authenController';
 import LoginView from './views/loginView';
+import STORAGE_KEYS from './constants/storageKeys';
+import LocalStorage from './utils/localStorage';
 
 const listNoteModel = new NoteModel();
 
@@ -14,17 +16,11 @@ const listNoteController = new NoteController(listNoteModel, listNoteView);
 const authenController = new AuthenController(loginView);
 
 (() => {
-  const page = document.body.className;
+  const localStorage = new LocalStorage();
 
-  switch (page) {
-    case 'index-page':
-      authenController.init();
-      break;
-    case 'home-page':
-      listNoteController.init();
-      break;
-    default:
-      console.log('Oops! Have something went wrong');
-      break;
+  if (!localStorage.getItems(STORAGE_KEYS.IS_USER_LOGGED_IN)) {
+    authenController.init();
+  } else {
+    listNoteController.init();
   }
 })();
