+++ javascript/big-practice/src/scripts/views/loginView.js
@@ -1,7 +1,7 @@
 import { selectDOMById, selectDOMClass } from '../utils/querySelectDOM';
 import { ERROR_MESSAGE } from '../constants/message';
 import { hideError, showError } from '../utils/handleError';
-import user from '../constants/mockUser';
+import user from '../../../data/mockUser';
 import LocalStorage from '../utils/localStorage';
 import STORAGE_KEYS from '../constants/storageKeys';
